+++ kamehouse-shell/bin/kamehouse/install-kamehouse-groot.sh
@@ -1,9 +1,5 @@
 #!/bin/bash
 
-# Execute from the root of the kamehouse git project:
-# chmod a+x ./kamehouse-shell/bin/kamehouse/install-kamehouse-groot.sh
-# ./kamehouse-shell/bin/kamehouse/install-kamehouse-groot.sh
-
 COL_BLUE="\033[1;34m"
 COL_BOLD="\033[1m"
 COL_CYAN="\033[1;36m"
@@ -14,21 +10,26 @@ COL_RED="\033[1;31m"
 COL_YELLOW="\033[1;33m"
 COL_MESSAGE=${COL_GREEN}
 
+KAMEHOUSE_USER=""
+
 main() {
   parseArguments "$@"
   log.info "Installing ${COL_PURPLE}kamehouse-groot${COL_MESSAGE}"
-  log.info "User running this script needs ${COL_RED}sudo ln${COL_MESSAGE} permissions"
   createApacheHttpdUserSymLink
   log.info "Done installing ${COL_PURPLE}kamehouse-groot!"
 }
 
 createApacheHttpdUserSymLink() {
-  local USERNAME=`whoami`
   log.info "Creating symlink on /var/www for apache user www-data to access kamehouse-shell scripts"
+  if (( $EUID != 0 )); then
+    log.error "This script needs to run as root. Exiting..."
+    exit 1
+  fi
   log.info "This step is necessary to execute successfully kamehouse-shell commands from groot"
   log.info "This script is only necessary on ${COL_PURPLE}linux${COL_MESSAGE}. It's not needed in ${COL_PURPLE}windows" 
-  sudo ln -s /home/${USERNAME}/programs /var/www
-  log.info "Ignore ${COL_PURPLE}sudo${COL_MESSAGE} error on windows. Run with ${COL_PURPLE}sudo${COL_MESSAGE} permissions on linux"
+  ln -s /home/${KAMEHOUSE_USER}/programs /var/www
+  mkdir -p /var/www/logs
+  chmod a+rwx /var/www/logs
 }
 
 log.info() {
@@ -44,11 +45,14 @@ log.error() {
 }
 
 parseArguments() {
-  while getopts ":h" OPT; do
+  while getopts ":hu:" OPT; do
     case $OPT in
     ("h")
       printHelp
       exit 0
+      ;;
+    ("u")
+      KAMEHOUSE_USER=$OPTARG
       ;;
     (\?)
       log.error "Invalid argument $OPTARG"
@@ -56,6 +60,12 @@ parseArguments() {
       ;;
     esac
   done
+
+  if [ -z "${KAMEHOUSE_USER}" ]; then
+    log.error "Option -u is required"
+    printHelp
+    exit 1
+  fi
 }
 
 printHelp() {
@@ -64,6 +74,7 @@ printHelp() {
   echo -e ""
   echo -e "  Options:"  
   echo -e "     ${COL_BLUE}-h${COL_NORMAL} display help"
+  echo -e "     ${COL_BLUE}-u (username)${COL_NORMAL} user running kamehouse [${COL_RED}required${COL_NORMAL}]"
 }
 
 main "$@"