+++ kamehouse-shell/bin/kamehouse/install-kamehouse-shell-root.sh
@@ -1,9 +1,5 @@
 #!/bin/bash
 
-# Execute from the root of the kamehouse git project:
-# chmod a+x ./kamehouse-shell/bin/kamehouse/install-kamehouse-shell-root.sh
-# ./kamehouse-shell/bin/kamehouse/install-kamehouse-shell-root.sh
-
 COL_BLUE="\033[1;34m"
 COL_BOLD="\033[1m"
 COL_CYAN="\033[1;36m"
@@ -14,16 +10,16 @@ COL_RED="\033[1;31m"
 COL_YELLOW="\033[1;33m"
 COL_MESSAGE=${COL_GREEN}
 
+KAMEHOUSE_USER=""
+
 main() {
   parseArguments "$@"
   log.info "Setting up root user for kamehouse"
-  log.info "Run this script as the user who installed and runs kamehouse"
-  if (( $EUID == 0 )); then
-    log.error "Running this script as root. It needs to run as a normal user. Exiting..."
+  if (( $EUID != 0 )); then
+    log.error "This script needs to run as root. Exiting..."
     exit 1
   fi
-  log.info "User running this script needs ${COL_RED}sudo su${COL_MESSAGE} permissions"
-  sudo su -c "cd ${HOME}/git/kamehouse ; ./kamehouse-shell/bin/kamehouse/install-kamehouse-shell.sh"
+  cd /home/${KAMEHOUSE_USER}/git/kamehouse ; ./kamehouse-shell/bin/kamehouse/install-kamehouse-shell.sh
   log.info "To ${COL_RED}uninstall${COL_MESSAGE} kamehouse-shell for root, run as root ${COL_PURPLE}cd ${HOME}/git/kamehouse ; ./scripts/uninstall-kamehouse.sh"
   log.info "Finished setting up root user for kamehouse"
 }
@@ -41,11 +37,14 @@ log.error() {
 }
 
 parseArguments() {
-  while getopts ":hop" OPT; do
+  while getopts ":hu:" OPT; do
     case $OPT in
     ("h")
       printHelp
       exit 0
+      ;;
+    ("u")
+      KAMEHOUSE_USER=$OPTARG
       ;;
     (\?)
       log.error "Invalid argument $OPTARG"
@@ -53,6 +52,12 @@ parseArguments() {
       ;;
     esac
   done
+
+  if [ -z "${KAMEHOUSE_USER}" ]; then
+    log.error "Option -u is required"
+    printHelp
+    exit 1
+  fi
 }
 
 printHelp() {
@@ -61,6 +66,7 @@ printHelp() {
   echo -e ""
   echo -e "  Options:"  
   echo -e "     ${COL_BLUE}-h${COL_NORMAL} display help"
+  echo -e "     ${COL_BLUE}-u (username)${COL_NORMAL} user running kamehouse [${COL_RED}required${COL_NORMAL}]"
 }
 
 main "$@"