+++ kamehouse-shell/bin/lin/kamehouse/httpd-status.sh
@@ -1,10 +1,20 @@
 #!/bin/bash
 
+if (( $EUID == 0 )); then
+  HOME="/var/www"
+fi
+
 # Import common functions
 source ${HOME}/programs/kamehouse-shell/bin/common/common-functions.sh
 if [ "$?" != "0" ]; then
 	echo -e "\033[1;36m$(date +%Y-%m-%d' '%H:%M:%S)\033[0;39m - [\033[1;31mERROR\033[0;39m] - \033[1;31mAn error occurred importing common-functions.sh\033[0;39m"
 	exit 1
+fi
+# Import kamehouse functions
+source ${HOME}/programs/kamehouse-shell/bin/common/kamehouse/kamehouse-functions.sh
+if [ "$?" != "0" ]; then
+  echo -e "\033[1;36m$(date +%Y-%m-%d' '%H:%M:%S)\033[0;39m - [\033[1;31mERROR\033[0;39m] - \033[1;31mAn error occurred importing kamehouse-functions.sh\033[0;39m"
+  exit 1
 fi
 
 LOG_PROCESS_TO_FILE=true
@@ -13,9 +23,8 @@ HTTPD_PORT=""
 
 mainProcess() {
   log.info "Searching for apache httpd process"
-  log.warn "User running this script needs ${COL_RED}sudo netstat${COL_DEFAULT_LOG} permissions"
-  sudo netstat -nltp | grep ${HTTPD_PORT} | grep apache 
-  HTTPD_PID=`sudo netstat -nltp | grep ${HTTPD_PORT} | grep apache | awk '{print $7}' | cut -d '/' -f 1`
+  setSudoKameHouseCommand "netstat -nltp"
+  HTTPD_PID=`${SUDO_KAMEHOUSE_COMMAND} | grep ${HTTPD_PORT} | grep apache | awk '{print $7}' | cut -d '/' -f 1`
   if [ -z ${HTTPD_PID} ]; then
     log.info "Apache httpd is not running"
   else