+++ kamehouse-shell/bin/lin/kamehouse/httpd-stop.sh
@@ -1,20 +1,28 @@
 #!/bin/bash
 
+if (( $EUID == 0 )); then
+  HOME="/var/www"
+fi
+
 # Import common functions
 source ${HOME}/programs/kamehouse-shell/bin/common/common-functions.sh
 if [ "$?" != "0" ]; then
 	echo -e "\033[1;36m$(date +%Y-%m-%d' '%H:%M:%S)\033[0;39m - [\033[1;31mERROR\033[0;39m] - \033[1;31mAn error occurred importing common-functions.sh\033[0;39m"
 	exit 1
 fi
+# Import kamehouse functions
+source ${HOME}/programs/kamehouse-shell/bin/common/kamehouse/kamehouse-functions.sh
+if [ "$?" != "0" ]; then
+  echo -e "\033[1;36m$(date +%Y-%m-%d' '%H:%M:%S)\033[0;39m - [\033[1;31mERROR\033[0;39m] - \033[1;31mAn error occurred importing kamehouse-functions.sh\033[0;39m"
+  exit 1
+fi
 
 LOG_PROCESS_TO_FILE=true
-DEFAULT_HTTPD_PORT=80
-HTTPD_PORT=""
 
 mainProcess() {
   log.info "Stopping apache httpd server"
-  log.warn "User running this script needs ${COL_RED}sudo service${COL_DEFAULT_LOG} permissions"
-  sudo service apache2 stop
+  setSudoKameHouseCommand "service apache2 stop"
+  ${SUDO_KAMEHOUSE_COMMAND}
 }
 
 main "$@"